### 2023-03-13 #5249 @editorialbot
[PRE REVIEW]: Baargin: a Nextflow workflow for the automatic analysis of bacterial genomics data with a focus on Antimicrobial Resistance
**Submitting author:** @jhayer (<a href="http://orcid.org/0000-0003-4899-9637">Juliette Hayer</a>)
**Repository:** https://github.com/jhayer/baargin
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
