### 2023-03-13 #5249 @Kevin-Mattheus-Moerman
@jhayer I am the AEiC for this track and here to process initial steps. I have just flagged this submission for a scope review by our editorial board. This is normal for submissions of this size. The scope review should be completed within about 2 weeks and will determined if this submission conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).
