+++ paper/paper.bib
@@ -89,7 +89,7 @@ year = {2017}
 @article{Wood:2019,
 abstract = {Although Kraken's k-mer-based approach provides a fast taxonomic classification of metagenomic sequence data, its large memory requirements can be limiting for some applications. Kraken 2 improves upon Kraken 1 by reducing memory usage by 85%, allowing greater amounts of reference genomic data to be used, while maintaining high accuracy and increasing speed fivefold. Kraken 2 also introduces a translated search mode, providing increased sensitivity in viral metagenomics analysis.},
 author = {Wood, Derrick E. and Lu, Jennifer and Langmead, Ben},
-doi = {10.1186/S13059-019-1891-0/FIGURES/2},
+doi = {10.1186/S13059-019-1891-0},
 file = {:Users/2021jh002/Library/Application Support/Mendeley Desktop/Downloaded/Wood, Lu, Langmead - 2019 - Improved metagenomic analysis with Kraken 2.pdf:pdf},
 issn = {1474760X},
 journal = {Genome Biology},
