### 2023-04-18 b3ab885
changing reads input handling strategy for better management of input files patterns
