### 2023-04-21 #5249 @mberacochea
> @coughls @mberacochea @rcannood - Would you like to review this submission to JOSS? Our reviews are checklist-driven, meaning that we provide you with a checklist and you verify that the submission meets each of the specified criteria. Please let me know if you're available to review. Thank you!

I'm happy to review this one :+1:
