### 2023-04-22 #5397 @editorialbot
[REVIEW]: Baargin: a Nextflow workflow for the automatic analysis of bacterial genomics data with a focus on Antimicrobial Resistance
**Submitting author:** @jhayer (<a href="http://orcid.org/0000-0003-4899-9637">Juliette Hayer</a>)
**Repository:** https://github.com/jhayer/baargin
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
