### 2023-05-09 jhayer/baargin#13 @rcannood
[JOSS] @rcannood's feedback for the JOSS submission
I'm currently reviewing this repository's JOSS submission (https://github.com/openjournals/joss-reviews/issues/5397). I'll update this issue throughout the review process and will notify you when I'm finished with the review of the first version of the manuscript :)

Robrecht

-----

- [x] #34

- [x] #44
  nextflow run jhayer/baargin -r 0.1.0 --help
  ```

- [x] #19

## Installation instructions

- [x] #35

## Functionality documentation

- [x] #21

## Automated tests

- [x] #47

- [x] #24

# Software paper

- [x] #14

- [x] #29

## Quality of writing

- [x] #18
- [x] #16
