### 2023-05-11 jhayer/baargin#13 @rcannood
Alright, I went through the checklist. @jhayer if you disagree with any of the comments, feel free to discuss.

It would be fantastic if you could tackle each comment in this issue through a separate PR (whenever possible), because that makes my life so much easier in checking whether the changes has resolved my concerns ;)
