### 2023-05-17 0362f81
Fixed summary in paper.md (fix #14)
