### 2023-05-17 fd8b961
Statement of need paper.md (fix #16 and #17)
I have rephrased some parts in the paragraph Statement of need. Is is easier to read and understand now.
