### 2023-05-21 a32408f
Clearer parameters documentation (fix #21)
I have tried to make clearer and more structured description of the parameters available for the workflow. Let me know if it needs more modifications
