+++ README.md
@@ -175,14 +175,18 @@ nextflow run /path/to/baargin/main.nf -profile docker \
 ```
 
 
-## Parameters
+# Parameters
 
-For running the workflow needs 3 main parameters:
+## Mandatory
+
+For running the workflow you need 3 mandatory parameters:
 1. the input datasets: 3 possible inputs:
-  - directory containing paired-end short reads (Illumina type)
-  - directory containing already assembled contigs/scaffolds
+  - directory containing paired-end short reads (Illumina type): path to provide with the parameter `--illumina`
+OR
+  - directory containing already assembled contigs/scaffolds: path to provide with the parameter `--contigs`
+OR
   - an index CSV file indicating path to short reads and long reads; for hybrid input requiring Unicycler hybrid assembly.
-  The CSV index file should look as below and must include the columns headers:
+  The CSV index file is provided with the parameter `--hybrid_index ` and should look as below and must include the columns headers:
 
 ```
 sampleID,read1,read2,ont
@@ -190,10 +194,83 @@ sampleID,read1,read2,ont
 365,test_illu_hybrid/365_1.fq,test_illu_hybrid/365_2.fq,test_ont/barcode01_concat.fastq
 ```
 
-2. Three mandatory databases should already be in the `db` directory within the `baargin` directory (these paths are default in the `nextflow.config` after you have run the `download_db.py` script). To overwrite in the command line if different
+2. Three mandatory databases should already be in the `db` directory within the `baargin` directory (these paths are set by default in the `nextflow.config` after you have run the `download_db.py` script). 
+
+
+Note: If you wish to set a different path for these 3 DB, you can overwrite in the command line using the parameters:
+`--card_db path/to/card/db`
+`--kraken2_db path/to/kraken/db`
+`--plasmidfinder_db path/to/plasmidfinder/db`
 
 3. A TaxID (NCBI Taxonomy ID) to which extract from to get "decontaminated" scaffolds/contigs belonging to the expected bacterial taxon. It can be a TaxID corresponding to an *order*, a *genus* or a *species*, and all the contigs classified by Kraken2 under this specified taxon and lower in the taxonomy (children taxa) will be retrieved as decontaminated.
+This parameter is provided as follow:
+`--species_taxid "562"` (in this example we want to extract seqeunces classified as Taxonomy_ID 562, which corresponds to *E. coli*
 
+
+You set the output directory with `--output path/outputdir`, it is set by default to `./results` in the `nextflow.config` file.
+
+You set the temporary directory with `--tmpdir path/tmpdir`, it is set by default to `./tmpdir` in the `nextflow.config` file.
+
+You set the NextFlow work directory with `-work-dir path/work`, it is set by default to `./work` in the `nextflow.config` file.
+
+## Optional parameters
+
+**A. The databases**
+
+You can set the paths to the optional databases (for amrfinder, bakta and platon) if you have them in local using the following parameters:
+
+```
+--amrfinder_db "/path/to/local/databases/amrfinder/latest"
+--bakta_db "/path/to/local/databases/bakta_db"
+--platon_db "/path/to/local/databases/platon/db"
+```
+
+*Note:* a database for amrfinder is included in its container.
+
+If you want to avoid the download of Busco databases every time you run the pipeline, you can specify a local BUSCO database using:
+
+`--busco_db_offline "/path/to/local/databases/busco_downloads"`
+
+**B. The taxonomy parameters**
+
+Additionally to the mandatory `--species_taxid` parameter, you can add the following parameters:
+
+```
+--genus = "Escherichia"
+--species = "coli"
+```
+
+For BUSCO and AMRFinder, you can specify specific datasets to use.
+
+```
[…]
