+++ .github/workflows/ci.yml
@@ -1,0 +1,37 @@
+name: ci
+
+# Controls when the action will run. 
+# Triggers the workflow on push or pull request events
+on:
+  push:
+    branches:
+      - main
+      - ci
+  pull_request:
+      - main
+
+permissions:
+  contents: write
+
+jobs:
+  # This workflow contains a single job called "bu
+  test:
+    runs-on: continuumio/miniconda3
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: 3.x
+      - uses: actions/cache@v2
+        with:
+          key: ${{ github.ref }}
+          path: .cache
+      - name: Install baargin conda env
+        run: conda env create -f conda_environment_baargin.yml
+       - name: Activate baargin env
+        run: conda activate baargin
+      - name: Run baargin test pipeline
+        run: nextflow run main.nf -profile docker \ 
+             --illumina '/path/to/baargin/data' --genus 'Escherichia' --species 'coli' \
+             --busco_lineage 'enterobacterales_odb10' --amrfinder_organism 'Escherichia' \
+             --species_taxid '562' --output './results_test' -resume
