+++ .github/workflows/ci.yml
@@ -7,8 +7,7 @@ on:
     branches:
       - main
       - ci
-  pull_request:
-      - main
+  pull_request
 
 permissions:
   contents: write
