+++ .github/workflows/ci.yml
@@ -7,7 +7,9 @@ on:
     branches:
       - main
       - ci
-  pull_request
+  pull_request:
+    types:
+      - opened
 
 permissions:
   contents: write
