### 2023-05-22 870743e
update README for illumina input: folder+pattern params
