### 2023-05-26 cb08207
add contributing guidelines to README.md (fix #24)
