### 2023-05-26 jhayer/baargin#26 @mberacochea
[JOSS] @mberacochea's feedback for the JOSS submission
This ticket is to gather my review comments for the JOSS submission (https://github.com/openjournals/joss-reviews/issues/5397).

- [x] #36
  - Added a comment -> https://github.com/jhayer/baargin/issues/36
- [x] #41

TODO: I will run the pipeline once the installation instructions have been amended.

## A statement of need
- [x] #27
  - README greatly improved -> https://github.com/jhayer/baargin/pull/22
## Installation instructions
- [x] #37
  - README greatly improved, with better installation instructions -> https://github.com/jhayer/baargin/pull/22
## Automated tests
- [x] #48
- [x] #28
  - Added -> https://github.com/jhayer/baargin/pull/25/files

# Software paper
## State of field
- [x] #30
  - Added -> https://github.com/jhayer/baargin/commit/dd2e39fac8a4214a12fbe4432ccf4fd1d836e583
## Quality of writing
- [x] Line 61: instead of " The contigs are taxonomically assigned using Kraken2.." -> "Taxonomic assignment of the contigs is performed using Kraken2..."
- [x] Line 67: instead of "A quality check of the assembly is achieved by Quast..." -> "A quality check of the assembly is conducted using Quast...".
- [x] Line 73: "The contigs are then submitted to a plasmid identification with PlasmidFinder" -> "The contigs are subsequently submitted to plasmid identification using PlasmidFinder."
- [x] Line 98: " that belong the expected taxon" -> "that belong to the expected taxon"
- [x] Nextflow is written as "NextFlow" and "Nextflow" throughout the text.

# Other comments
Overall the nextflow code is correct and easy to follow. I would recommend to split the pipeline into several sub-workflows, that helps with the readability.

- [x] #31
  - Fixed -> https://github.com/jhayer/baargin/pull/33
- [x] #38
   - Fixed https://github.com/jhayer/baargin/pull/39
- [x] #32
  - Fixed -> https://github.com/jhayer/baargin/pull/33
- [x] #42
  - Fixed -> https://github.com/jhayer/baargin/pull/43
- [x] #45
   - Fixed -> https://github.com/jhayer/baargin/pull/46]
