### 2023-05-28 jhayer/baargin#26 @mberacochea
OK, my first review is completed now. @jhayer please, if you disagree with any of my comments please let me know :).
