### 2023-05-28 #5397 @mberacochea
Hey, my inital review is done -> https://github.com/jhayer/baargin/issues/26
