+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: 'Baargin: a Nextflow workflow for the automatic analysis of bacterial
 genomics data with a focus on Antimicrobial Resistance'
 tags:
-  - NextFlow
+  - Nextflow
   - Whole Genome Shotgun
   - Genomics
   - Long reads sequencing technology
@@ -49,12 +49,12 @@ at a time. As a counterpart, these experiments produce large amount of data that
 needs to be analysed by various bioinformatics methods and tools for reconstructing
 the genomes and therefore identify their specific features and the genetic
 determinants of the AMR. For automating the bioinformatics analysis of multiple
-strains, we have developed a NextFlow [@DITommaso:2017] workflow called *baargin*
-(Bacterial Assembly and Antimicrobial Resistance Genes detection In NextFlow)
+strains, we have developed a Nextflow [@DITommaso:2017] workflow called *baargin*
+(Bacterial Assembly and Antimicrobial Resistance Genes detection In Nextflow)
 [https://github.com/jhayer/baargin](https://github.com/jhayer/baargin).
 It enables to conduct sequencing reads quality control, genome assembly and annotation,
 Multi-Locus Sequence Typing and plasmid identification, as well as antimicrobial
-resistance determinants detection, and pangenome analysis. The use of NextFlow,
+resistance determinants detection, and pangenome analysis. The use of Nextflow,
 a workflow management system, makes our workflow portable, flexible, and able to
 conduct reproducible analyses.
 
@@ -108,18 +108,18 @@ Fastp [@Chen:2018].
 3. De novo assembly is run using SPAdes [@Prjibelski:2020] if only short reads were
 provided, and with Unicycler [@Wick:2017] for a hybrid assembly when short and
 long reads are provided.
-4. The contigs are taxonomically assigned using Kraken2 [@Wood:2019] and the contigs classified at
+4. Taxonomic assignment of the contigs is performed using Kraken2 [@Wood:2019] and the contigs classified at
 the taxonomic level provided by the user (with the taxid, and including the children taxa)
 are retrieved and therefore named as *"deconta"* for decontaminated contigs [@Lu:2022]. The dataset
 containing all the contigs are named as *"raw"*. From here all the steps except the
 annotation (8) will be performed on both sets of contigs *"raw"* and *"deconta"*.
-5. A quality check of the assembly is achieved by Quast [@Gurevich:2013] and BUSCO [@Manni:2021].
+5. A quality check of the assembly is conducted using Quast [@Gurevich:2013] and BUSCO [@Manni:2021].
 For BUSCO, the users have the possibility to specify the taxonomic lineage database
 to use for searching the housekeeping genes (at the class level of the strain to
 analyse for example: *enterobacterales_odb10*)
 6. The contigs (*raw* and *deconta*) are then screened to identify the sequence type of
 the strain using MLST tool (Multi-Locus Sequence Typing) [@Seemann:2022].
-7. The contigs are then submitted to a plasmid identification with PlasmidFinder [@Carattoli:2014]
+7. The contigs are subsequently submitted to plasmid identification using PlasmidFinder [@Carattoli:2014]
 and additionally with Platon if the user provides a database for it [@Schwengers:2020].
 8. Antimicrobial Resistance Genes (ARGs) are then searched in the contigs using both CARD RGI [@Alcock:2023]
 and the NCBI AMRFinderPlus [@Feldgarden:2021]. For certain species only, AMRFinderPlus
@@ -151,7 +151,7 @@ We presented here an easy-to-use workflow for Bacterial Assembly and Antimicrobi
 Resistance Genes detection In Nextflow: *baargin*. It allows the users to analyse
 genomic datasets from short and long sequencing reads, of several bacterial strains from
[…]
