### 2023-06-12 906876d
improve installation procedure, fix #34, fix #35, fix #36, fix 37
