+++ README.md
@@ -1,27 +1,12 @@
 # baargin: Bacterial Assembly and Antimicrobial Resistance Genes detection In NextFlow
+=========================================
+<img align="right" src="doc/img/IRD.png" width="200" height="66" /> <img align="right" src="doc/img/MIVEGEC.png" width="100" height="66" />
 Workflow for analysis of Whole Genome Sequencing (WGS) data with AntiMicrobial Resistance (AMR) focus
 
-The *baargin* workflow allows to perform a complete *in silico* analysis of bacterial genomics datasets from multiple isolates in parallel.
-The user can input sequencing datasets from short reads only or from both short and long reads (hybrid assembly Illumina + Oxford Nanopore Technologies), of several bacterial strains from the same species in one command line. The workflow will automatically assemble the genomes, assign the taxonomy of the assembled sequences (contigs) and specifically extract the sequences that belong the expected taxon (specified by the user). The contigs extracted at that step are then called *deconta* (for decontaminated).
-For both *raw* and *deconta* contigs, *baargin* will then:
--	identify their sequence type (MLST),
--	screen for plasmids sequences,
--	detect Antimicrobial Resistance Genes (ARGs) and mutations, and 
--	perform genome annotation (*deconta* only). 
-
-The workflow compiles and summarizes the results from all the analysis steps, allowing comparative studies. As a last step, *baargin* performs a pangenome analysis of all the genomics datasets provided, producing the basis for the construction of a phylogenetic tree. 
-A flowchart below describes the process.
-
-
-
-
-<img src="doc/img/IRD.png" width="300" height="100" /> <img src="doc/img/MIVEGEC.png" width="150" height="100" />
-
-<img src="doc/img/baargin_flowchart.jpg" width="900" height="500" />
 
 ## Table of Contents
 
-   * [Prerequisites](#prerequisites)
+   * [Foreword](#Foreword)
    * [Installation](#installation)
       * [Using conda](#using-conda)
       * [Old school - Manually](#old-school---manually)
@@ -35,104 +20,173 @@ A flowchart below describes the process.
    * [Citation](#citation)
    * [Author](#author-and-contributors)
 
+## Foreword
 
-## Prerequisites
+The *baargin* workflow allows to perform a complete *in silico* analysis of bacterial genomics datasets from multiple isolates in parallel.
+The user can input sequencing datasets from short reads only or from both short and long reads (hybrid assembly Illumina + Oxford Nanopore Technologies), of several bacterial strains from the same species in one command line. The workflow will automatically assemble the genomes, assign the taxonomy of the assembled sequences (contigs) and specifically extract the sequences that belong the expected taxon (specified by the user). The contigs extracted at that step are then called *deconta* (for decontaminated).
+For both *raw* and *deconta* contigs, *baargin* will then:
+-	identify their sequence type (MLST),
+-	screen for plasmids sequences,
+-	detect Antimicrobial Resistance Genes (ARGs) and mutations, and 
+-	perform genome annotation (*deconta* only). 
 
-You need to have installed Docker or Singularity as the workflow uses containers to run the different tools.
+The workflow compiles and summarizes the results from all the analysis steps, allowing comparative studies. As a last step, *baargin* performs a pangenome analysis of all the genomics datasets provided, producing the basis for the construction of a phylogenetic tree. 
+A flowchart below describes the process.
+
+
+<img src="doc/img/baargin_flowchart.jpg" width="900" height="500" />
 
 
 ## Installation
 
-### Using conda
+The prerequisites to run the pipeline are:  
[…]
