### 2023-06-12 605e774
add singu and docker installation
