### 2023-06-12 7228c40
fix path + add info
