### 2023-06-14 cb63087
use csv module for tabular formatting
