### 2023-06-15 #5397 @jhayer
Hi, We addressed most of your comments by  several PRs (linked to issues numbers).
We would like to merge all the PRs before proceeding with the remaining issues (tests: issues #47 and #48, and release: #44).
Could you please @rcannood and @mberacochea check that all this is ok before we merge?
Thank you !
