### 2023-07-06 #5397 @fboehm
Thanks for the replies, @rcannood and @mberacochea ! And thanks for agreeing to review the updates. As you proceed, please feel free to discuss things here. Also, feel free to open issues in the baargin repo, like you did with the first iteration of reviews.

Please know that I'm here to help, so if anything is unclear, don't hesitate to ask me about it. Thanks again!
