### 2023-07-09 jhayer/baargin#36 @mberacochea
Hi! Thank you for the changes; I think they are a great improvement.

I understand the usefulness of https://github.com/jhayer/baargin/blob/main/download_db.py, but this is the main reason for the requirement of a Conda environment. By also providing a set of curl or wget commands, it would be possible to make the Conda environment not required. The most technically savvy users will probably notice this. I feel that not having a Conda environment makes using the pipeline more accessible, especially for Nextflow users.

For plasmidfinder_db , it's possible to download the repo using the Bibucket Download Repo URL: https://bitbucket.org/genomicepidemiology/plasmidfinder_db/get/314d85f43e4e.zip.

Something like:
```bash
mkdir dbs/plasmidfinder
cd dbs/plasmidfinder
wget https://bitbucket.org/genomicepidemiology/plasmidfinder_db/get/314d85f43e4e.zip -O plasmidfinderdb && unzip -j plasmidfinderdb.zip
```
