### 2023-07-09 #5397 @mberacochea
Hi @fboehm and @jhayer. I have updated my PR, and I've added one minor comment about the installation -> https://github.com/jhayer/baargin/issues/36.

I will go through the installation and functional testing as soon as the PRs are merged
