### 2023-07-14 #5397 @rcannood
Thanks for the updates! It's great that you split up the different items into different PRs, which allowed me to go through them and quickly find out whether the PR addressed the concerns I raised earlier.

With the updated codebase I managed to go through the documentation and execute the pipeline myself. Most of the checkboxes have now been ticked. I updated jhayer/baargin#13 based on some minor issues I encountered.
