### 2023-07-24 jhayer/baargin#36 @jhayer
Hi ! Thank you for the comments.
We had actually added a "manually" section to the documentation for the databases download (using wget, curl or git clone).
