### 2023-08-28 jhayer/baargin#26 @jhayer
Hi @mberacochea,
We now finally went through all your comments and fixed everything. Thank you for your valuable input.
