### 2023-09-11 jhayer/baargin#26 @mberacochea
Well.. I was able to run the test data set. I think this PR can be closed :)
