### 2023-09-17 #5397 @rcannood
Hi all! Sorry for taking so long to get back to this.

I'm thrilled with the changes that the authors have made! The revised manuscript looks great; I'm very happy that the workflow doesn't generate unnecessary warnings anymore; and the test profile provided by @jhayer works great!

Don't forget to update the manuscript, because the latest build still contains the previously mentioned typos (e.g. "NextFlow" → "Nextflow" and "Hight" → "High"), but I assume @editorialbot will take care of that automatically.

With the completion of jhayer/baargin#13, I can now recommend this manuscript.
