### 2023-09-17 jhayer/baargin#13 @rcannood
Thanks @jhayer ! I feel like all of my concerns have been addressed :+1:
