### 2023-09-20 #5397 @jhayer
It was already fixed so I could not see the typos in my latest build (strange). I have double checked in the paper.md and it is ok there, so it should be fine for the final build. Thanks to all of you !
