### 2023-09-25 #5397 @jhayer
Hi @fboehm - how long time should we expect for the article to be accepted and available online?
