### 2023-09-25 #5397 @fboehm
Thanks for the question, @jhayer. The next steps require several actions from me and a few tasks from you. In a moment, I'll create the checklists for the remaining tasks. Once they are completed, I'll tell the EiCs that the submission is ready for publication. I expect that we'll need at least a week from now.
