### 2023-09-28 #5397 @jhayer
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
baargin-v1.0.0

> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
https://doi.org/10.5281/zenodo.8386399

> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.
