### 2023-10-18 #5397 @fboehm
> @jhayer I am the EiC on this track and here to help with final steps. I have checked this review, the paper, the repository, and the archive. Most seems in order. I do have the below points which need your attention.
>
> * [ ]  Please amend the version tag listed on your archive link to match `v1.0.0`, it currently says `v1`
> * [ ]  Your text features British English (e.g. `analysed` and `customise`) as well as American English (e.g.  `standardize`, `parallelize`, and `summarize`), please amend so you consistently use one.
> * [ ]  Check if `...using MLST tool...` should be rewritten as `...using the MLST tool...`
> * [ ]  Currently you have cited a URL using the following .bib file entry:
>
> ```
> @misc{Seemann:2022,
> author = {Seemann, Torsten},
> booktitle = {https://github.com/tseemann/mlst},
> mendeley-groups = {Chile_ESBLEcoli},
> title = {{MLST: https://github.com/tseemann/mlst}},
> year = {2022}
> }
> ```
>
> I would recommend that you alter it to have the URL in a `url` field so it renders properly in the references (note I also fixed the title).
>
> ```
> @misc{Seemann:2022,
> author = {Seemann, Torsten},
> title = {mlst},
> year = {2022},
> url = "https://github.com/tseemann/mlst"
> }
> ```

Thanks, @Kevin-Mattheus-Moerman, for spotting these issues. I'll be sure to look for similar issues in future submissions.
