### 2023-10-19 a2e0950
fix am. eng. to british eng., clean bib file
