+++ paper/paper.bib
@@ -17,7 +17,6 @@ file = {:Users/2021jh002/Library/Application Support/Mendeley Desktop/Downloaded
 issn = {1546-1696},
 journal = {Nature Biotechnology 2017 35:4},
 keywords = {Computational biology and bioinformatics,Data publication and archiving},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {apr},
 number = {4},
 pages = {316--319},
@@ -47,7 +46,6 @@ doi = {10.1093/BIOINFORMATICS/BTY560},
 issn = {1367-4803},
 journal = {Bioinformatics},
 keywords = {Contact,chen@haploxcom},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {sep},
 number = {17},
 pages = {i884--i890},
@@ -66,7 +64,6 @@ doi = {10.1002/CPBI.102},
 issn = {1934-340X},
 journal = {Current Protocols in Bioinformatics},
 keywords = {biosynthetic gene clusters,de novo assembly,genome assembly,metagenomes,plasmids,transcriptome},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {jun},
 number = {1},
 pages = {e102},
@@ -85,7 +82,6 @@ doi = {10.1371/JOURNAL.PCBI.1005595},
 issn = {1553-7358},
 journal = {PLOS Computational Biology},
 keywords = {Bacterial genomics,Chromosome pairs,Genome sequencing,Genomics,Klebsiella pneumoniae,Operons,Plasmid construction,Sequence alignment},
-mendeley-groups = {Pipeline_nf_wgs},
 month = {jun},
 number = {6},
 pages = {e1005595},
@@ -105,7 +101,6 @@ file = {:Users/2021jh002/Library/Application Support/Mendeley Desktop/Downloaded
 issn = {1474760X},
 journal = {Genome Biology},
 keywords = {Alignment-free methods,Metagenomics,Metagenomics classification,Microbiome,Minimizers,Probabilistic data structures},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {nov},
 number = {1},
 pages = {1--13},
@@ -124,7 +119,6 @@ doi = {10.1038/S41596-022-00738-Y},
 file = {:Users/2021jh002/Library/Application Support/Mendeley Desktop/Downloaded/Lu et al. - 2022 - Metagenome analysis using the Kraken software suite.pdf:pdf},
 issn = {17502799},
 journal = {Nature protocols},
-mendeley-groups = {Pipeline_nf_wgs},
 month = {dec},
 number = {12},
 pages = {2815},
@@ -143,7 +137,6 @@ doi = {10.1093/BIOINFORMATICS/BTT086},
 file = {:Users/2021jh002/Library/Application Support/Mendeley Desktop/Downloaded/Gurevich et al. - 2013 - QUAST quality assessment tool for genome assemblies.pdf:pdf},
 issn = {1367-4803},
 journal = {Bioinformatics},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {apr},
 number = {8},
 pages = {1072--1075},
@@ -166,7 +159,6 @@ isbn = {002162155.1},
 issn = {15371719},
 journal = {Molecular Biology and Evolution},
 keywords = {completeness,eukaryotes,genome,metagenomes,microbes,prokaryotes,quality assessment,transcriptome,viruses},
-mendeley-groups = {Chile_ESBLEcoli,Pipeline_nf_wgs},
 month = {sep},
 number = {10},
 pages = {4647--4654},
[…]
